was redesign header of desktop large
