+++ src/pages/HomePage/sections/Header/Header.jsx
@@ -1,12 +1,32 @@
-import { Link } from 'react-router';
+import { Link as RouterLink, Link } from 'react-router';
 import { useTranslation } from 'react-i18next';
-import logo from '../../assets/logo.svg';
-import linkdin from '../../assets/linkdin.svg';
+import navigationsLinks from '@router/links.js';
+import { openModal } from '@redux/slices/modal/modalSlice.js';
+import { modalNames } from '@utils/constants/modalNames.js';
+import { useDispatch, useSelector } from 'react-redux';
 import mail from '../../assets/mail.svg';
+import linkdin from '../../assets/linkdin.svg';
+import logo from '../../assets/logo.svg';
 import styles from './Header.module.scss';
 
 const Header = () => {
   const { t } = useTranslation();
+  const dispatch = useDispatch();
+  const { isAuth } = useSelector((state) => state.tokens);
+  const handleOpenLogin = () => dispatch(openModal({ modalType: modalNames.loginModal }));
+  const myProfile = () => {
+    if (!isAuth)
+      return (
+        <button className={`btn btn-secondary btn-xl ${styles.login}`} onClick={handleOpenLogin}>
+          <span>{t('home.header.login')}</span>
+        </button>
+      );
+    return (
+      <Link className={`btn btn-secondary btn-xl ${styles.login}`} component={RouterLink} to={navigationsLinks.profile}>
+        {t('home.header.profile')}
+      </Link>
+    );
+  };
 
   return (
     <header className={styles.header__bg} id='home'>
@@ -32,14 +52,16 @@ const Header = () => {
               </li>
             </ul>
           </nav>
-
-          <div className={styles.header__socialIcons}>
-            <a aria-label='LinkedIn' href='https://www.linkedin.com'>
-              <img alt='LinkedIn Icon' src={linkdin} />
-            </a>
-            <a aria-label='Email' className={styles.header__socialIcons__dn} href='mailto:info@skillzzy.com'>
-              <img alt='Mail Icon' src={mail} />
-            </a>
+          <div className={styles.header__social_login}>
+            <div className={styles.header__socialIcons}>
+              <a aria-label='LinkedIn' href='https://www.linkedin.com'>
+                <img alt='LinkedIn Icon' src={linkdin} />
+              </a>
+              <a aria-label='Email' className={styles.header__socialIcons__dn} href='mailto:info@skillzzy.com'>
+                <img alt='Mail Icon' src={mail} />
+              </a>
+            </div>
+            {myProfile()}
           </div>
         </div>
       </div>
