+++ src/pages/HomePage/sections/HeroSection/HeroSection.module.scss
@@ -248,23 +248,8 @@
     }
   }
 
-  .login {
-    max-width: 220px;
-    width: 50%;
-
-    @include desktop {
-      max-width: 187px;
-    }
-
-    @include tablet-large {
-      width: 50%;
-      max-width: 100%;
-    }
-  }
-
   @include desktop {
-    .registration,
-    .login {
+    .registration {
       display: flex;
       align-items: center;
       font-size: 18px !important;
@@ -274,8 +259,7 @@
   }
 
   @include tablet {
-    .registration,
-    .login {
+    .registration {
       display: none;
     }
   }
