+++ src/utils/locales/uk/index.js
@@ -8,8 +8,10 @@ export const uk = {
       header: {
         howItWorks: 'Як це працює',
         aboutUs: 'Про нас',
-        faq: 'Часті питання',
+        faq: 'FAQ',
         contacts: 'Контакти',
+        login: 'Увійти',
+        profile: 'Профіль',
       },
       hero: {
         title: 'Мультифункціональна платформа від розробників для розробників',
@@ -25,7 +27,6 @@ export const uk = {
           'Опануйте навички проходження співбесід за допомогою реальної практики разом з експертами та однодумцями. Отримуйте персоналізований зворотний зв’язок, об’єктивні оцінки та глибокі аналітичні інсайти, щоб упевнено підготуватися до успіху на будь-якій співбесіді!',
         secondaryTextMobile: "Щоб зареєструватися, відкрийте сайт на комп'ютері",
         registration: 'Реєстрація',
-        login: 'Увійти',
       },
       howItWorks: {
         titleText1: 'Як це',
